# apis/rest/openapi-v1.yml
openapi: 3.0.0
info: 
  title:	users API
  description: Master cloud apps TFM users API.
  version: 1.0.0
servers:
  - url: https://localhost:8443/api/v1
  - url: https://users-tfm-dev-amartinm82.cloud.okteto.net/api/v1
paths:
  /admins/auth:
    post:
      description: Authenticate an admin using passed credentials and return a JWT token.
      tags:
        - admins
      operationId: adminAuthentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: a JSON web token (JWT).
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
                required:
                  - token
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/InvalidCredentialsError'
        '500':
          $ref: '#/components/responses/InternalError'
  /admins:
    post:
      description: Creates an admin and returns the admin ID.
      tags:
        - admins
      operationId: createAdmin
      requestBody:
        required: true
        description: A JSON object that contains the username (email) and password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created admin
          headers:
            Location:
              required: true
              description: URL to get the newly-created admin.
              schema:
                type: string
                example: https://localhost:8443/v1/admins/1
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: ID of the created admin.
                    example: 1
                required:
                  - id
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: Username already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Username already exists.
        '500':
          $ref: '#/components/responses/InternalError'

  /auth:
    post:
      description: Authenticate an user using passed credentials and return a JWT token.
      tags:
        - users
      operationId: authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: a JSON web token (JWT).
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
                required:
                  - token
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/InvalidCredentialsError'
        '500':
          $ref: '#/components/responses/InternalError'
  /users:
    post:
      description: Creates an user and returns the user ID.
      tags:
        - users
      operationId: createUser
      requestBody:
        required: true
        description: A JSON object that contains the username (email) and password.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Created user
          headers:
            Location:
              required: true
              description: URL to get the newly-created user.
              schema:
                type: string
                example: https://localhost:8443/v1/users/1
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: ID of the created user.
                    example: 1
                required:
                  - id
          links:
            GetUserByUserId:   
              operationId: getUserById
              parameters:
                userId: '$response.body#/id'
              description: >
                The `id` value returned in the response can be used as
                the `userId` parameter in `GET /users/{userId}`.
            AddBalanceToUser:   
              operationId: addBalance
              parameters:
                userId: '$response.body#/id'
              description: >
                The `id` value returned in the response can be used as
                the `userId` parameter in `POST /users/{userId}/balance`.
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: Username already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Username already exists.
        '500':
          $ref: '#/components/responses/InternalError'
          
  /users/{userId}:  
    get:
      description: Get user info by ID.
      tags:
        - users
      operationId: getUserById
      parameters:
        - in: path
          description: Numeric identifier of the user to get.
          name: userId
          required: true
          schema:
            type: integer
          example: 1
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/NotAllowedError'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: User not found.
        '500':
          $ref: '#/components/responses/InternalError'
          
  /users/{userId}/balance:  
    post:
      description: Add balance to user.
      tags:
        - users
      operationId: addBalance
      parameters:
        - in: path
          description: Numeric identifier of the user to get.
          name: userId
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        description: A JSON object that contains the amount to add to user balance.
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  description: amount to add to user balance.
                  type: number
                  minimum: 0
                  exclusiveMinimum: true
                  example: 35.80
              required:
                - amount
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User info.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/NotAllowedError'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: User not found.
        '500':
          $ref: '#/components/responses/InternalError'

components:
  responses:
    InvalidCredentialsError:
      description: Invalid credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid credentials.
    UnauthorizedError:
      description: Access token is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Access token is missing or invalid.
    NotAllowedError:
      description: Logged user doesn't have permission to access the resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: You don't have permissions to access this resource.
    BadRequestError:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad request.
    InternalError:
      description: Internal error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal error.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          description: user email
          type: string
          format: email
          example: a.martinmar.2020@alumnos.urjc.es 
        password:
          description: User passwod (UpperCase, LowerCase and Number) with at least 8 characters.
          type: string
          format: password
          minLength: 8
          pattern: ^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?!.*\s).*$
          example: P4ssword
    UserResponse:
      type: object
      properties:
        id:
          description: user ID.
          type: integer
          example: 1
        username:
          description: user email.
          type: string
          format: email
          example: a.martinmar.2020@alumnos.urjc.es 
        balance:
          description: user available balance.
          type: number
          example: 35.80
      required:
        - id
        - username
        - balance
    Error:
        type: object
        description: error message.
        properties:
          error:
            type: string
        required: 
          - error